refactor: rename typescript/server to typescript/language
