+++ src/decoration/service.ts
@@ -5,7 +5,7 @@ import { TypeScriptParser } from '@/parser'
 import type { SymbolKind } from '@/symbol'
 import { Logger } from '@/logger'
 import type { SymbolColorMap } from '@/theme'
-import { TypeScriptLanguageService, TypeScriptServerProbe } from '@/typescript/server'
+import { TypeScriptLanguageService, TypeScriptServerProbe } from '@/typescript/language'
 import { SymbolResolver } from './resolver'
 import type { ResolveTarget } from './resolver'
 import type { DocumentCache, SymbolOccurrence } from './types'
