+++ src/di/container.ts
@@ -1,8 +1,8 @@
 import 'reflect-metadata'
 import { Container } from 'inversify'
 import { TOKENS } from './tokens'
-import { TypeScriptLanguageService } from '@/typescript/server'
-import { TypeScriptServerProbe } from '@/typescript/server'
+import { TypeScriptLanguageService } from '@/typescript/language'
+import { TypeScriptServerProbe } from '@/typescript/language'
 import { TypeScriptParser } from '@/parser'
 import { ThemeColorResolver } from '@/theme'
 import { DecorationService } from '@/decoration'
