+++ src/symbol/resolvers/plugin.test.ts
@@ -4,7 +4,7 @@ import { TypeScriptServerNotLoadedError } from '@/symbol/errors'
 import { PluginSymbolResolver } from './plugin'
 import { SymbolKind } from '@/symbol/types'
 import { RESPONSE_KEY, type PluginResponse, type ResolveResponse } from '@/typescript/plugin/protocol'
-import type { TypeScriptLanguageService, DefinitionResult } from '@/typescript/server'
+import type { TypeScriptLanguageService, DefinitionResult } from '@/typescript/language'
 
 const ALL_FALSE: Omit<ResolveResponse, 'id'> = {
   isFunction: false,
