+++ src/symbol/resolvers/semanticToken.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, beforeEach, vi } from 'vitest'
 import * as vscode from 'vscode'
 import { SemanticTokenSymbolResolver } from './semanticToken'
 import { SymbolKind } from '@/symbol/types'
-import type { TypeScriptLanguageService, DefinitionResult } from '@/typescript/server'
+import type { TypeScriptLanguageService, DefinitionResult } from '@/typescript/language'
 
 function createMockDocument(uri = 'file:///test.ts') {
   return { uri: vscode.Uri.parse(uri) } as unknown as vscode.TextDocument
