+++ src/typescript/language/probe.ts
